refactor: extracting metric utils functions into a module.
